chore: move classes to new testdata package
